accessible/tests/mochitest/tree/test_dockids.html

<!DOCTYPE html>
<html>
<head>
  <title>Test document hierarchy</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>

  <script type="application/javascript">
  function doTest()
  {
    var tree =
     { DOCUMENT: [
       { PARAGRAPH: [ // head
         { PARAGRAPH: [ // link
           { STATICTEXT: [] }, // generated content
           { STATICTEXT: [] } // generated content
         ] }
       ] },
       { TEXT_LEAF: [ ] }, // body text
       { ENTRY: [ ] }, // input under document element
       { PARAGRAPH: [ // link under document element
         { TEXT_LEAF: [ ] }, // link content
         { STATICTEXT: [ ] }, // generated content
         { STATICTEXT: [ ] } // generated content
       ] },
       { LINK: [ // anchor under document element
         { TEXT_LEAF: [ ] } // anchor content
       ] },
     ] };
    testAccessibleTree(getNode("iframe").contentDocument, tree);

    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  addA11yLoadEvent(doTest);
  </script>
</head>

<body>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=608887"
     title="Elements appended outside the body aren't accessible">
    Mozilla Bug 608887
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <iframe src="dockids.html" id="iframe"></iframe>
</body>
</html>
